Delete
View changelog for details.
